kernelgen/pyproject.toml
[project]
name = "kernelgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==25.1.0",
    "beautifulsoup4==4.14.3",
    "boto3==1.40.61",
    "chz>=0.4.0",
    "datasets>=4.5.0",
    "inspect-ai==0.3.180",
    "matplotlib==3.10.8",
    "modal>=1.3.3",
    "nest-asyncio>=1.6.0",
    "nltk==3.9.2",
    "openai==2.21.0",
    "pytest==9.0.2",
    "python-dotenv==1.2.1",
    "scipy==1.17.0",
    "textarena==0.7.4",
    "textual==8.0.0",
    "tiktoken==0.12.0",
    "tinker>=0.13.1",
    "torch>=2.10.0",
    "torchvision==0.25.0",
    "transformers==5.2.0",
    "wandb>=0.25.0",
]

[tool.uv.sources]
tinker = { git = "https://github.com/thinking-machines-lab/tinker.git", rev = "a187cb5b2f2ad0f3eef6b93eab2770edcf0f01c7" }
